move related data structures to download events loop mod
